Getting group now has a new reducer
